requirements.txt
appdirs==1.4.4
    # via black
appnope==0.1.2
    # via ipython
attrs==21.2.0
    # via
    #   -r requirements.in
    #   cattrs
    #   pytest
backcall==0.2.0
    # via ipython
black==21.5b1
    # via -r requirements.in
cachetools==4.2.2
    # via google-auth
cattrs==1.6.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==8.0.1
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colored==1.4.2
    # via -r requirements.in
coverage[toml]==5.5
    # via pytest-cov
decorator==5.0.9
    # via ipython
flake8==3.9.2
    # via -r requirements.in
google-auth-oauthlib==0.4.4
    # via
    #   -r requirements.in
    #   gspread
google-auth==1.30.1
    # via
    #   google-auth-oauthlib
    #   gspread
gspread==3.7.0
    # via -r requirements.in
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
ipdb==0.13.7
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
ipython==7.23.1
    # via ipdb
isort==5.8.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via -r requirements.in
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.9
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.1.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest-cov==2.12.0
    # via -r requirements.in
pytest-mock==3.6.1
    # via -r requirements.in
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytz==2021.1
    # via -r requirements.in
regex==2021.4.4
    # via black
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via requests-oauthlib
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via google-auth
toml==0.10.2
    # via
    #   black
    #   coverage
    #   ipdb
    #   pep517
    #   pytest
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typed-ast==1.4.3
    # via mypy
typing-extensions==3.10.0.0
    # via mypy
urllib3==1.26.4
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
